backend/requirements.txt
Django>=4.2.11
django-cors-headers>=4.3.1
djangorestframework>=3.16.0
boto3>=1.34.0
PyJWT>=2.8.0 